A Novel QoS Prediction Approach for Cloud Service Based on Bayesian Networks Model
A Novel QoS Prediction Approach for Cloud Service Based on Bayesian Networks Model

In the dynamic, complex and changeable cloud computing environment, Quality Of Service (QoS) is an important basis for the selection of different cloud services. Therefore, the prediction of cloud services QoS can help users to choose the most suitable service at hand. The software and hardware and resources of three-layer structure for cloud computing will impact on cloud services QoS, but existing QoS prediction approaches are not consider the three-layer structure on the influence of the cloud service QoS.The CPU usage, physical memory usage and the number of processes of infrastructure layer have definitely influenced QoS. In order to address this limitation, in the paper, a Bayesian network model of QoS prediction for cloud services is proposed. Firstly, an initial and basic Bayesian network model is established by collecting data from the infrastructure layer, the platform layer and the application layer. Then the Bayesian network is trained and updated to obtain the cloud service QoS prediction model. Finally, a set of experiments based on collected data from the real cloud service environment has been conducted to validate the proposed approach. Experimental results show that the prediction approach is effective and accurate.

Data Description
Therefore, in order to build a good collection of data from the server cloud platform, the underlying data infrastructure layer and the platform layer is mainly related to the collection of hardware resources, namely the number of processes, CPU usage, physical memory usage, etc., at the application layer collection service response time pelagic resources and service availability and other data. Response time refers to the service consumer issues a message and receive a response to the time interval between milliseconds. Availability is defined as the service is available, is described as Ture express service is available, is described as False indicates that the service is unavailable.